+++ esp-metadata-generated/src/_build_script_utils.rs
@@ -3291,6 +3291,7 @@ impl Chip {
                     "soc_has_lp_apm",
                     "soc_has_lp_clkrst",
                     "soc_has_lpwr",
+                    "soc_has_lp_gpio",
                     "soc_has_lp_io_mux",
                     "soc_has_lp_peri",
                     "soc_has_lp_tee",
@@ -3333,6 +3334,7 @@ impl Chip {
                     "i2s_driver_supported",
                     "interrupts_driver_supported",
                     "io_mux_driver_supported",
+                    "lp_io_driver_supported",
                     "phy_driver_supported",
                     "psram_driver_supported",
                     "rng_driver_supported",
@@ -3419,6 +3421,7 @@ impl Chip {
                     "i2s_supports_pcm2pdm",
                     "interrupts_status_registers=\"3\"",
                     "interrupt_controller=\"clic\"",
+                    "lp_io_version=\"v4\"",
                     "phy_combo_module",
                     "psram_extmem_origin=\"1107296256\"",
                     "rng_apb_cycle_wait_num=\"16\"",
@@ -3505,6 +3508,7 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_lp_apm",
                     "cargo:rustc-cfg=soc_has_lp_clkrst",
                     "cargo:rustc-cfg=soc_has_lpwr",
+                    "cargo:rustc-cfg=soc_has_lp_gpio",
                     "cargo:rustc-cfg=soc_has_lp_io_mux",
                     "cargo:rustc-cfg=soc_has_lp_peri",
                     "cargo:rustc-cfg=soc_has_lp_tee",
@@ -3547,6 +3551,7 @@ impl Chip {
                     "cargo:rustc-cfg=i2s_driver_supported",
                     "cargo:rustc-cfg=interrupts_driver_supported",
                     "cargo:rustc-cfg=io_mux_driver_supported",
+                    "cargo:rustc-cfg=lp_io_driver_supported",
                     "cargo:rustc-cfg=phy_driver_supported",
                     "cargo:rustc-cfg=psram_driver_supported",
                     "cargo:rustc-cfg=rng_driver_supported",
@@ -3633,6 +3638,7 @@ impl Chip {
                     "cargo:rustc-cfg=i2s_supports_pcm2pdm",
                     "cargo:rustc-cfg=interrupts_status_registers=\"3\"",
                     "cargo:rustc-cfg=interrupt_controller=\"clic\"",
+                    "cargo:rustc-cfg=lp_io_version=\"v4\"",
                     "cargo:rustc-cfg=phy_combo_module",
                     "cargo:rustc-cfg=psram_extmem_origin=\"1107296256\"",
                     "cargo:rustc-cfg=rng_apb_cycle_wait_num=\"16\"",
